requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
pytest==8.2.0
pytest-django==4.8.0